requirements.txt
flask
websocket-client
flask-socketio
eventlet  
requests